requirements.txt
Flask
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
flask-cors
python-dotenv
gunicorn
